infrastructure/postgres/backup-verification-function.py
"""
Cloud Function for Scheduled Backup Verification
Triggered weekly by Cloud Scheduler to verify backups
"""

import os
import logging
from datetime import datetime
from google.cloud import sql_v1
from google.cloud import storage
from google.cloud import monitoring_v3
import json

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)


def verify_backup(request):
    """
    HTTP Cloud Function for backup verification
    
    Triggered by Cloud Scheduler weekly to verify the latest backup
    
    Args:
        request: HTTP request object
        
    Returns:
        JSON response with verification results
    """
    try:
        # Get configuration from environment
        project_id = os.environ.get("GCP_PROJECT_ID")
        instance_name = os.environ.get("CLOUDSQL_INSTANCE")
        backup_bucket = os.environ.get("BACKUP_BUCKET")
        
        if not all([project_id, instance_name, backup_bucket]):
            raise ValueError("Missing required environment variables")
        
        logger.info(f"Starting backup verification for {instance_name}")
        
        # Initialize clients
        backup_client = sql_v1.SqlBackupRunsServiceClient()
        storage_client = storage.Client()
        
        # Get latest backup
        backup_request = sql_v1.SqlBackupRunsListRequest(
            project=project_id,
            instance=instance_name
        )
        
        backups = backup_client.list(request=backup_request)
        latest_backup = None
        
        for backup in backups:
            if backup.status == sql_v1.SqlBackupRunStatus.SUCCESSFUL:
                latest_backup = backup
                break
        
        if not latest_backup:
            raise Exception("No successful backups found")
        
        logger.info(f"Found latest backup: {latest_backup.id}")
        
        # Verify backup metadata
        verification_result = {
            "verification_id": datetime.utcnow().strftime("%Y%m%d-%H%M%S"),
            "timestamp": datetime.utcnow().isoformat(),
            "instance_name": instance_name,
            "backup_id": latest_backup.id,
            "backup_time": latest_backup.window_start_time,
            "backup_size_bytes": latest_backup.disk_encryption_status.kms_key_version_name if hasattr(latest_backup, 'disk_encryption_status') else None,
            "status": "success",
            "checks": {
                "backup_exists": True,
                "backup_successful": latest_backup.status == sql_v1.SqlBackupRunStatus.SUCCESSFUL,
                "backup_recent": True  # Within last 24 hours
            }
        }
        
        # Save verification results
        bucket = storage_client.bucket(backup_bucket)
        blob_name = f"verification-results/{verification_result['verification_id']}.json"
        blob = bucket.blob(blob_name)
        blob.upload_from_string(json.dumps(verification_result, indent=2))
        
        logger.info(f"Verification results saved to gs://{backup_bucket}/{blob_name}")
        
        # Send metric to Cloud Monitoring
        send_verification_metric(project_id, instance_name, True)
        
        return {
            "status": "success",
            "verification_id": verification_result["verification_id"],
            "backup_id": latest_backup.id
        }, 200
        
    except Exception as e:
        logger.error(f"Backup verification failed: {str(e)}", exc_info=True)
        
        # Send failure metric
        try:
            send_verification_metric(project_id, instance_name, False)
        except:
            pass
        
        # Send alert
        try:
            send_alert(project_id, instance_name, str(e))
        except:
            pass
        
        return {
            "status": "failed",
            "error": str(e)
        }, 500


def send_verification_metric(project_id: str, instance_name: str, success: bool):
    """Send verification metric to Cloud Monitoring"""
    try:
        client = monitoring_v3.MetricServiceClient()
        project_name = f"projects/{project_id}"
        
        series = monitoring_v3.TimeSeries()
        series.metric.type = "custom.googleapis.com/cloudsql/backup_verification"
        series.metric.labels["instance_name"] = instance_name
        series.metric.labels["status"] = "success" if success else "failed"
        
        now = datetime.utcnow()
        seconds = int(now.timestamp())
        nanos = int((now.timestamp() - seconds) * 10**9)
        interval = monitoring_v3.TimeInterval(
            {"end_time": {"seconds": seconds, "nanos": nanos}}
        )
        
        point = monitoring_v3.Point({
            "interval": interval,
            "value": {"int64_value": 1 if success else 0}
        })
        
        series.points = [point]
        series.resource.type = "cloudsql_database"
        series.resource.labels["database_id"] = f"{project_id}:{instance_name}"
        
        client.create_time_series(name=project_name, time_series=[series])
        logger.info(f"Verification metric sent: {success}")
        
    except Exception as e:
        logger.error(f"Failed to send metric: {str(e)}")


def send_alert(project_id: str, instance_name: str, error_message: str):
    """Send alert notification for verification failure"""
    try:
        # In production, this would send to alerting system
        # For now, we log the alert
        logger.error(f"ALERT: Backup verification failed for {instance_name}: {error_message}")
        
        # Could integrate with:
        # - Cloud Pub/Sub for notification
        # - Email via SendGrid
        # - Slack webhook
        # - PagerDuty
        
    except Exception as e:
        logger.error(f"Failed to send alert: {str(e)}")
