infrastructure/postgres/backup-verification-requirements.txt
# Requirements for backup verification Cloud Function

google-cloud-sql==1.7.0
google-cloud-storage==2.10.0
google-cloud-monitoring==2.15.1
functions-framework==3.4.0
